feat(network): add an HttpSend interceptor entry point for Ktor
`ktorClientPlugin()` can only be applied inside an `HttpClient { }` config
block, so a client built by a DI container or a library could not be
inspected without changing its construction site.

Add `ktorSendInterceptor(client)`, which returns an `HttpSendInterceptor`
the caller registers on an already-built client:

    client.plugin(HttpSend).intercept(agent.ktorSendInterceptor(client))

This mirrors the OkHttp adapter, which returns an `Interceptor` for the
caller to pass to `addInterceptor`. Both Ktor entry points stay supported:
`install()` deduplicates by `AttributeKey` and reads declaratively, so it
remains the better choice when the construction site is reachable.

The two share one mechanism — the `Send` hook is implemented as
`client.plugin(HttpSend).intercept` — so the round-trip logic moves to
`monitorSend()`, parameterised only by how the send continues (`proceed`
vs `execute`). Behaviour is unchanged; the existing tests cover that.

`client` is a parameter because mocked responses synthesize an
`HttpClientCall`, and `Sender` does not expose the client it sends through.

Tests pin the ordering invariants both entry points rely on: install order
determines nesting, `HttpRedirect` always sits outside the monitor so every
hop is its own transaction, and duplicate registration deduplicates for
`install()` but double-records for the interceptor.
